support for 1.21.3

+++ src/main/java/net/pitan76/itemalchemy/tile/EMCBatteryTile.java
@@ -23,10 +23,10 @@ import net.pitan76.mcpitanlib.api.event.nbt.ReadNbtArgs;
 import net.pitan76.mcpitanlib.api.event.nbt.WriteNbtArgs;
 import net.pitan76.mcpitanlib.api.event.tile.TileTickEvent;
 import net.pitan76.mcpitanlib.api.gui.args.CreateMenuEvent;
-import net.pitan76.mcpitanlib.api.gui.inventory.sided.args.AvailableSlotsArgs;
-import net.pitan76.mcpitanlib.api.gui.v2.ExtendedScreenHandlerFactory;
 import net.pitan76.mcpitanlib.api.gui.inventory.IInventory;
 import net.pitan76.mcpitanlib.api.gui.inventory.sided.VanillaStyleSidedInventory;
+import net.pitan76.mcpitanlib.api.gui.inventory.sided.args.AvailableSlotsArgs;
+import net.pitan76.mcpitanlib.api.gui.v2.ExtendedScreenHandlerFactory;
 import net.pitan76.mcpitanlib.api.network.PacketByteUtil;
 import net.pitan76.mcpitanlib.api.network.v2.ServerNetworking;
 import net.pitan76.mcpitanlib.api.tile.ExtendBlockEntityTicker;
@@ -90,7 +90,7 @@ public class EMCBatteryTile extends EMCStorageBlockEntity implements ExtendBlock
 
         //if (maxEMC <= storedEMC) return;
 
-        coolDown += (int) (1 + Math.pow(maxEMC / 10000 - 1, 2));
+        coolDown += (int) (1 + Math.pow((double) maxEMC / 10000 - 1, 2));
         if (coolDown >= getMaxCoolDown())
             coolDown = 0;
